Move components to files
