+++ packages/vkui/src/components/ScrollSaver/ScrollSaver.tsx
@@ -1,12 +1,10 @@
 import * as React from 'react';
-import { useNavId } from '../../components/NavIdContext/useNavId';
-import { useNavDirection } from '../../components/NavTransitionDirectionContext/NavTransitionDirectionContext';
 import { usePatchChildrenRef } from '../../hooks/usePatchChildrenRef';
-import { useIsomorphicLayoutEffect } from '../../lib/useIsomorphicLayoutEffect';
 import { HasRootRef } from '../../types';
-import { useScrollSaverCache } from './ScrollSaverContext';
+import { ScrollSaveModeType } from './types';
+import { useScrollSaver } from './useScrollSaver';
 
-interface ScrollSaverProps {
+export interface ScrollSaverProps {
   /*
    * Уникальный идентификатор элемента скролл которого надо запомнить.
    * Важно задавать id, так как на одной панели может понадобится запомнить позиции нескольких скролл-боксов.
@@ -17,7 +15,7 @@ interface ScrollSaverProps {
    * `forward` - позиция скролла сохраняется только при переходе вперёд и восстанавливается при переходе назад.
    * `always` - позиция скролла сохраняется и при переходе вперёд и при переходе назад.
    **/
-  saveMode?: 'forward' | 'always';
+  saveMode?: ScrollSaveModeType;
   /*
    * Если передан реакт-компонент, то он должен поддерживать getRootRef.
    **/
@@ -34,82 +32,9 @@ interface ScrollSaverProps {
  *
  * @see https://vkcom.github.io/VKUI/#/ScrollSaver
  */
-export function ScrollSaver(props: ScrollSaverProps) {
+export const ScrollSaver = (props: ScrollSaverProps) => {
   const [childrenRef, children] = usePatchChildrenRef(props.children, props.useGetRef);
   useScrollSaver(childrenRef, props.id, props.saveMode);
 
   return children;
-}
-
-export function useScrollSaver<T extends HTMLElement>(
-  elementRef: React.RefObject<T>,
-  id: ScrollSaverProps['id'],
-  saveMode: ScrollSaverProps['saveMode'] = 'forward',
-) {
-  const uniqueId = useUniqueId(id);
-  const direction = useNavDirection();
-  const scrollSaverCache = useScrollSaverCache();
-
-  useIsomorphicLayoutEffect(
-    function handleScrollPosition() {
-      const scrollId = uniqueId;
-      const refNode = elementRef.current;
-
-      function restoreScrollPosition() {
-        if (!refNode) {
-          return;
-        }
-
-        const scrollPosition = scrollSaverCache[scrollId];
-        if (!scrollPosition) {
-          return;
-        }
-
-        const shouldRestoreMovingBackwards = direction === 'backwards' && saveMode === 'forward';
-        const shouldRestoreMovingForwards = saveMode === 'always';
-        const shouldRestore = shouldRestoreMovingBackwards || shouldRestoreMovingForwards;
-        if (!shouldRestore) {
-          return;
-        }
-
-        const { inlineStart, blockStart } = scrollSaverCache[scrollId];
-        refNode.scrollLeft = inlineStart;
-        refNode.scrollTop = blockStart;
-      }
-
-      restoreScrollPosition();
-
-      return function saveScrollPositionOnUnmount() {
-        if (!refNode) {
-          return;
-        }
-        scrollSaverCache[scrollId] = {
-          inlineStart: refNode.scrollLeft,
-          blockStart: refNode.scrollTop,
-        };
-      };
-    },
-    [direction, uniqueId, saveMode, elementRef],
-  );
-
-  return elementRef;
-}
-
-interface ScrollSaverWithoutChildrenProps<T = HTMLElement>
-  extends Omit<ScrollSaverProps, 'useGetRef'> {
-  elementRef: React.RefObject<T>;
-}
-
-/* Компонентный Вариант useScrollSaver хука для динамического рендеринга, чтобы можно было пробросить и использовать любой ref
- * children проп можно передать, но ref из него браться не будет */
-export function ScrollSaverWithoutChildren(props: ScrollSaverWithoutChildrenProps) {
-  useScrollSaver(props.elementRef, props.id, props.saveMode);
-
-  return props.children;
-}
-
-function useUniqueId(id: string) {
-  const { view: viewId, panel: panelId } = useNavId();
-  const uniqueId = `${viewId}-${panelId}-${id}`;
-  return uniqueId;
-}
+};
