+++ packages/vkui/src/components/ScrollSaver/types.ts
@@ -1,0 +1,1 @@
+export type ScrollSaveModeType = 'forward' | 'always';
